Fix snapshotting error handling
